chore: add a pre-commit config
If any developer uses [pre-commit](https://pre-commit.com/) it reliably
ensures that writing non-conventional commit messages do not happen, and
that the testsuite passes.

I also threw in two goodies of preventing TODO and FIXME occurances in
newly added lines (and yes, I had to git commit --no-verify to actually
make this commit).

One word of caveat: these are of cause only local tests, and are not
enforcable in any way, since they run on the committers machine.
Enforcement still needs to happen in the CI.