fix: make up a default value for version
If the version is not set correctly, the output becomes confusing.

So it's a bit better to disable this check when main.version was not set
correctly during build - for example when not using the Makefile to
build - which happens frequently when developing using an not fully
configured vscode instance.