chore: replace DOCKER_USERNAME secret by a variable
I personally find it annoying that this public piece of information is
redacted in the action output - GitHub tries to be on the safe side, so
it replaces every occurence of the value (`mstreubing`) everywhere in
the output. But since the value is not actually a secret value, it is
better to replace it by a variable.