fix(test): support running our test suite under `-trimpath`, closes #397 (#439)
* chore(deps): upgrade go-snaps to v0.5.10 that supports trimpath builds

* build(test): start running a pass of our testsuite under -trimpath

This commit is guaranteed to be broken, since our code not yet handles
the caveats regarding the current working dir assumption that go-snaps
now has.

* test: echo the commands being executed, to help the user find out which test phase fails

* fix(test): ensure snaps finds the snapshot directory by using an absolute path

Since we "tamper" with the current working directory to achieve a
certain output from files.GetRelativePath(), we now need to help
snaps find the __snapshots__ directory.