feat: ensure regex checks also work in more complex setups, e.g. mono-repos
The `^` character forces files to be at the root of a directory tree in order to
be excluded, however, in case of mono-repos for example, these files may be
nested deeper in the structure and wouldn't have been excluded despite being
e.g. lockfiles outside of user control.

See https://github.com/editorconfig-checker/editorconfig-checker/pull/462#issuecomment-2840244252